Refactor simulated state computers
